+++ apis/rws_devices_registry.cpp
@@ -498,7 +498,7 @@ bool deviceListMatchesHashRobonomics(Robonomics* robonomics) {
 	if (expected.length() == 0) {
 		return false;
 	}
-#if defined(DEBUG)
+#if defined(ALTRUIST_BUILD_DEBUG)
 	if (hash != expected) {
 		debug_outln_verbose(F("[RWS] registration hash mismatch"));
 		debug_outln_verbose(F("[RWS] expected: "), expected);